Updated /ChatShop
The help command now shows a brief remark about using "hand" for items
and "-" for prices.
It also now ensures that /buy and /sell are always the first two
commands shown, and that /csadmin is always the last command shown and
will only show if the user has permissions to use it.