EFind Improvements and Named Items Support
ChatShop now honors and stores the names of enchanted items. This
version of chatshop breaks with existing database schemas and will
require a database update, for which a script has been provided.

- /efind now allows search with "hand", whereupon the enchanted item in
hand supplies all of its enchantments as search parameters.
- /efind now interprets enchantments with no specified level to mean any
level of that enchant.
- /efind now allows "/efind #<lot>" in addition to "/efind <lot>".
- /efind now shows the provided enchantment parameters in the results
header.
- /find now redirects to /efind if the user searches "hand" and is
holding an enchanted item.

- /esell now detects and blocks items with enchantments higher than
level 9 which would break the ChatShop.

- Fixed a bug where enchanted books were not being correctly compared,
and always the first book in inventory was removed when posting the book
in hand which could be thusly duplicated.